Implement comprehensive anti-fingerprinting capabilities
- Add anti_fingerprint flag to BrowserConfig, BrowserContextConfig, and AgentSettings
- Create separate JavaScript files for basic and advanced anti-fingerprinting
- Implement canvas, WebGL, and audio fingerprinting protection
- Add dynamic platform values and screen property normalization
- Create comprehensive tests for anti-fingerprinting features
- Update documentation with usage examples and feature descriptions
