+++ browser_use/browser/scripts/__init__.py
@@ -1,0 +1,3 @@
+"""
+JavaScript scripts for browser anti-fingerprinting and other functionality.
+"""
